+++ libs/designer-ui/src/lib/editor/base/utils/helper.ts
@@ -4,7 +4,7 @@ import { ValueSegmentType } from '../../models/parameter';
 import { $isTokenNode } from '../nodes/tokenNode';
 import { guid } from '@microsoft/utils-logic-apps';
 import type { ElementNode } from 'lexical';
-import { $getNodeByKey, $isElementNode, $isTextNode } from 'lexical';
+import { $getNodeByKey, $isElementNode, $isLineBreakNode, $isTextNode } from 'lexical';
 
 export const removeFirstAndLast = (segments: ValueSegment[], removeFirst?: string, removeLast?: string): ValueSegment[] => {
   const n = segments.length - 1;
@@ -50,6 +50,52 @@ export const getChildrenNodes = (node: ElementNode, nodeMap?: Map<string, ValueS
     } else if ($isTokenNode(childNode)) {
       text += childNode.toString();
       nodeMap?.set(childNode.toString(), childNode.convertToSegment());
+    }
+    return text;
+  });
+  return text;
+};
+
+// interpolate tokens if they have not been interpolated
+export const getChildrenNodesWithTokenInterpolation = (node: ElementNode, nodeMap?: Map<string, ValueSegment>): string => {
+  let text = '';
+  let lastNodeWasText = '';
+  let lastNodeWasToken = '';
+  let numberOfDoubleQuotes = 0;
+  let numberOfQuotesAdded = 0;
+  node.getChildren().forEach((child) => {
+    const childNode = $getNodeByKey(child.getKey());
+    if (childNode && $isElementNode(childNode)) {
+      return (text += getChildrenNodesWithTokenInterpolation(childNode, nodeMap));
+    }
+    if ($isTextNode(childNode)) {
+      const childNodeText = childNode.__text.trim();
+      if (childNodeText.includes('"')) {
+        numberOfQuotesAdded = 0; // reset, the interpolation will be added with childNode
+      }
+      const missingAClosingInterpolation = numberOfQuotesAdded !== 0 && numberOfQuotesAdded % 2 === 1;
+      if (lastNodeWasToken === childNode.__prev && missingAClosingInterpolation) {
+        text += `"`;
+      }
+      text += childNodeText;
+      lastNodeWasText = childNode.__key;
+    } else if ($isTokenNode(childNode)) {
+      numberOfDoubleQuotes = (text.replace(/\\"/g, '').match(/"/g) || []).length;
+      if (lastNodeWasText === childNode.__prev && numberOfDoubleQuotes % 2 !== 1) {
+        text += `"`;
+        numberOfQuotesAdded++;
+      }
+      text += childNode.toString();
+      nodeMap?.set(childNode.toString(), childNode.convertToSegment());
+      lastNodeWasToken = childNode.__key;
+    }
+    if ($isLineBreakNode(childNode)) {
+      if (lastNodeWasText === childNode.__prev) {
+        lastNodeWasText = childNode.__key;
+      }
+      if (lastNodeWasToken === childNode.__prev) {
+        lastNodeWasToken = childNode.__key;
+      }
     }
     return text;
   });